+++ src/engines/L3/launch.js
@@ -16,6 +16,10 @@
  */
 
 import { RendererMain } from '@lightningjs/renderer'
+import { WebGlCoreRenderer, SdfTextRenderer } from '@lightningjs/renderer/webgl'
+import { CanvasCoreRenderer, CanvasTextRenderer } from '@lightningjs/renderer/canvas'
+import { Inspector } from '@lightningjs/renderer/inspector'
+
 import { Log } from '../../lib/log.js'
 import { SCREEN_RESOLUTIONS } from '../../constants.js'
 import colors from '../../lib/colors/colors.js'
@@ -24,17 +28,14 @@ import shaderLoader from './shaderLoader.js'
 
 export let renderer = {}
 
+const renderEngine = (settings) => {
+  const renderMode = 'renderMode' in settings ? settings.renderMode : 'webgl'
+
+  if (renderMode === 'webgl') return WebGlCoreRenderer
+  if (renderMode === 'canvas') return CanvasCoreRenderer
+}
+
 export default (App, target, settings = {}) => {
-  if ('fontLoader' in settings) {
-    Log.warn(
-      `
-
-Starting version 0.9.0 of Blits, the Launch setting \`fontLoader\` is not supported / required anymore.
-You can remove the option from your \`src/index.js\`-file. And you can safely remove the file \`src/fontLoader.js\` from your project.
-      `
-    )
-  }
-
   renderer = new RendererMain(
     {
       appWidth: settings.w || 1920,
@@ -50,12 +51,13 @@ You can remove the option from your \`src/index.js\`-file. And you can safely re
           ? settings.webWorkersLimit
           : window.navigator.hardwareConcurrency || 2,
       clearColor: (settings.canvasColor && colors.normalize(settings.canvasColor)) || 0x00000000,
-      enableInspector: settings.inspector || false,
+      inspector: settings.inspector === true ? Inspector : undefined,
       boundsMargin: settings.viewportMargin || 0,
       // gpu memory limit, converted from mb to bytes - defaults to 200mb
       txMemByteThreshold:
         'gpuMemoryLimit' in settings ? settings.gpuMemoryLimit * 1024 * 1024 : 200 * 1024 * 1024,
-      renderMode: 'renderMode' in settings ? settings.renderMode : 'webgl',
+      renderEngine: renderEngine(settings),
+      fontEngines: [SdfTextRenderer, CanvasTextRenderer],
     },
     target
   )